remove bottom sheets for pages
