+++ app/projects/[slug]/page.tsx
@@ -6,7 +6,6 @@ import { ChevronLeft, User, Calendar } from 'lucide-react'
 import { getProjectBySlug, getProjects } from '@/lib/projects'
 import { notFound } from 'next/navigation'
 import ReadingProgress from '@/components/ui/reading-progress'
-import ModalWrapper from '@/components/utility/ModalWrapper'
 import { Badge } from '@/components/ui/badge'
 
 
@@ -41,6 +40,7 @@ export default async function Project({
             className='object-cover'
             fill
             priority={true}
+            sizes="(max-width: 768px) 100vw, 768px"
           />
         </div>
       )}
@@ -91,23 +91,21 @@ export default async function Project({
   )
 
   return (
-    <ModalWrapper parentPath="/projects" isActive={true}>
-      <section className='pb-16 md:pb-24 pt-4 md:pt-32'>
-        <div className='container max-w-3xl'>
-          <Link
-            href='/projects'
-            className='mb-8 md:inline-flex hidden items-center gap-2 text-sm font-light text-muted-foreground transition-colors hover:text-foreground'
-          >
-            <ChevronLeft className='h-5 w-5' />
-            <span>Projekte</span>
-          </Link>
-          <div className="hidden md:block">
-            <ReadingProgress />
-          </div>
-          
-          {projectContent}
+    <section className='pb-16 pt-28 md:pb-24 md:pt-32'>
+      <div className='container max-w-3xl'>
+        <Link
+          href='/projects'
+          className='mb-8 inline-flex items-center gap-2 text-sm font-light text-muted-foreground transition-colors hover:text-foreground'
+        >
+          <ChevronLeft className='h-5 w-5' />
+          <span>Projekte</span>
+        </Link>
+        <div className="hidden md:block">
+          <ReadingProgress />
         </div>
-      </section>
-    </ModalWrapper>
+        
+        {projectContent}
+      </div>
+    </section>
   )
 }
